requirements.txt
flask==2.3.3
pygithub==2.3.0
mentat==1.0.8
python-dotenv==1.0.1
